+++ apps/meteor/client/views/account/preferences/PreferencesGlobalSection.tsx
@@ -1,5 +1,5 @@
 import type { SelectOption } from '@rocket.chat/fuselage';
-import { Accordion, Field, FieldGroup, FieldLabel, FieldRow, MultiSelect } from '@rocket.chat/fuselage';
+import { AccordionItem, Field, FieldGroup, FieldLabel, FieldRow, MultiSelect } from '@rocket.chat/fuselage';
 import { useUniqueId } from '@rocket.chat/fuselage-hooks';
 import { useUserPreference } from '@rocket.chat/ui-contexts';
 import React from 'react';
@@ -16,7 +16,7 @@ const PreferencesGlobalSection = () => {
 	const dontAskAgainListId = useUniqueId();
 
 	return (
-		<Accordion.Item title={t('Global')}>
+		<AccordionItem title={t('Global')}>
 			<FieldGroup>
 				<Field>
 					<FieldLabel htmlFor={dontAskAgainListId}>{t('Dont_ask_me_again_list')}</FieldLabel>
@@ -31,7 +31,7 @@ const PreferencesGlobalSection = () => {
 					</FieldRow>
 				</Field>
 			</FieldGroup>
-		</Accordion.Item>
+		</AccordionItem>
 	);
 };